requirements.txt
flask
torch
torchvision
numpy
pillow
python-dotenv
requests